[libcpu][riscv]移除ch32中的冗余文件,使用common下的文件
