+++ bsp/wch/risc-v/Libraries/Kconfig
@@ -14,7 +14,6 @@ config SOC_RISCV_SERIES_CH32V2
 config SOC_RISCV_SERIES_CH32V3
     bool
     select ARCH_RISCV
-    select ARCH_RISCV_FPU
     select SOC_RISCV_FAMILY_CH32
 
 config SOC_FAMILY_CH56X
