+++ bsp/wch/risc-v/Libraries/ch32v30x_libraries/bmsis/source/startup_ch32v30x.S
@@ -378,6 +378,10 @@ handle_reset:
  	la t0, _vector_base
     ori t0, t0, 3           
 	csrw mtvec, t0
+    
+    la t0, _eusrstack
+	addi t0, t0, -512
+    csrw mscratch,t0
 
     jal  SystemInit
 	la t0, entry
