+++ bsp/wch/risc-v/ch32v208w-r0/rtconfig.py
@@ -49,7 +49,7 @@ if PLATFORM == 'gcc':
     LFLAGS = DEVICE
     LFLAGS += ' -Wl,--gc-sections,-cref,-Map=' + MAP_FILE
     LFLAGS += ' -T ' + LINK_FILE
-
+    AFLAGS += ' -I.'
     CPATH = ''
     LPATH = ''
