+++ bsp/wch/risc-v/ch32v307v-r1/rtconfig.h
@@ -43,7 +43,7 @@
 #define RT_USING_CONSOLE
 #define RT_CONSOLEBUF_SIZE 128
 #define RT_CONSOLE_DEVICE_NAME "uart1"
-#define RT_VER_NUM 0x40101
+#define RT_VER_NUM 0x50000
 #define ARCH_RISCV
 
 /* RT-Thread Components */
@@ -69,6 +69,7 @@
 /* Device Drivers */
 
 #define RT_USING_DEVICE_IPC
+#define RT_UNAMED_PIPE_NUMBER 64
 #define RT_USING_SERIAL
 #define RT_USING_SERIAL_V1
 #define RT_SERIAL_RB_BUFSZ 64
@@ -156,11 +157,19 @@
 
 /* peripheral libraries and drivers */
 
+/* sensors drivers */
 
-/* kendryte-sdk: Kendryte SDK */
+
+/* touch drivers */
+
+
+/* Kendryte SDK */
 
 
 /* AI packages */
+
+
+/* Signal Processing and Control Algorithm Packages */
 
 
 /* miscellaneous packages */
@@ -172,6 +181,40 @@
 
 /* entertainment: terminal games and other interesting software packages */
 
+
+/* Arduino libraries */
+
+
+/* Projects */
+
+
+/* Sensors */
+
+
+/* Display */
+
+
+/* Timing */
+
+
+/* Data Processing */
+
+
+/* Data Storage */
+
+/* Communication */
+
+
+/* Device Control */
+
+
+/* Other */
+
+/* Signal IO */
+
+
+/* Uncategorized */
+
 #define SOC_RISCV_FAMILY_CH32
 #define SOC_RISCV_SERIES_CH32V3
 
@@ -181,10 +224,13 @@
 
 /* On-chip Peripheral Drivers */
 
+#define BSP_USING_GPIO
 #define BSP_USING_UART
 #define BSP_USING_UART1
+#define LSI_VALUE 40000
 
 /* Onboard Peripheral Drivers */
+
 
 /* Board extended module Drivers */
