+++ bsp/wch/risc-v/ch32v307v-r1/rtconfig.py
@@ -47,7 +47,7 @@ if PLATFORM == 'gcc':
     LFLAGS = DEVICE
     LFLAGS += ' -Wl,--gc-sections,-cref,-Map=' + MAP_FILE
     LFLAGS += ' -T ' + LINK_FILE
-
+    AFLAGS += ' -I.'
     CPATH = ''
     LPATH = ''
