+++ libcpu/risc-v/ch32/SConscript
@@ -7,10 +7,11 @@ Import('rtconfig')
 cwd     = GetCurrentDir()
 src     = Glob('*.c') + Glob('*.cpp')
 CPPPATH = [cwd]
+ASFLAGS = ' -I ' + cwd 
 
 if rtconfig.PLATFORM == 'gcc':
     src += Glob('*_gcc.S')
 
-group = DefineGroup('cpu', src, depend = [''], CPPPATH = CPPPATH)
+group = DefineGroup('cpu', src, depend = [''], CPPPATH = CPPPATH, ASFLAGS = ASFLAGS)
 
 Return('group')
