+++ libcpu/risc-v/common/cpuport.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2006-2021, RT-Thread Development Team
+ * Copyright (c) 2006-2023, RT-Thread Development Team
  *
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -7,6 +7,7 @@
  * Date           Author       Notes
  * 2018/10/28     Bernard      The unify RISC-V porting code.
  * 2020/11/20     BalanceTWK   Add FPU support
+ * 2023/01/04     WangShun     Adapt to CH32
  */
 
 #include <rthw.h>
@@ -125,7 +126,11 @@ rt_uint8_t *rt_hw_stack_init(void       *tentry,
     frame->epc     = (rt_ubase_t)tentry;
 
     /* force to machine mode(MPP=11) and set MPIE to 1 */
+    #ifdef ARCH_RISCV_FPU
     frame->mstatus = 0x00007880;
+    #else
+    frame->mstatus = 0x00001880;
+    #endif
 
     return stk;
 }
@@ -145,7 +150,9 @@ rt_weak void rt_hw_context_switch_interrupt(rt_ubase_t from, rt_ubase_t to, rt_t
 
     rt_interrupt_to_thread = to;
     rt_thread_switch_interrupt_flag = 1;
-
+#if defined(SOC_RISCV_FAMILY_CH32)
+    sw_setpend();
+#endif
     return ;
 }
 #endif /* end of RT_USING_SMP */
